+++ 3rd/nanovg/gl/nanovg_gl.h
@@ -56,41 +56,41 @@ enum NVGcreateFlags {
 
 #if defined NANOVG_GL2
 
-NVGcontext* nvgCreateGL2(int flags);
-void nvgDeleteGL2(NVGcontext* ctx);
+static NVGcontext* nvgCreateGL2(int flags);
+static void nvgDeleteGL2(NVGcontext* ctx);
 
-int nvglCreateImageFromHandleGL2(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
-GLuint nvglImageHandleGL2(NVGcontext* ctx, int image);
+static int nvglCreateImageFromHandleGL2(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
+static GLuint nvglImageHandleGL2(NVGcontext* ctx, int image);
 
 #endif
 
 #if defined NANOVG_GL3
 
-NVGcontext* nvgCreateGL3(int flags);
-void nvgDeleteGL3(NVGcontext* ctx);
+static NVGcontext* nvgCreateGL3(int flags);
+static void nvgDeleteGL3(NVGcontext* ctx);
 
-int nvglCreateImageFromHandleGL3(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
-GLuint nvglImageHandleGL3(NVGcontext* ctx, int image);
+static int nvglCreateImageFromHandleGL3(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
+static GLuint nvglImageHandleGL3(NVGcontext* ctx, int image);
 
 #endif
 
 #if defined NANOVG_GLES2
 
-NVGcontext* nvgCreateGLES2(int flags);
-void nvgDeleteGLES2(NVGcontext* ctx);
+static NVGcontext* nvgCreateGLES2(int flags);
+static void nvgDeleteGLES2(NVGcontext* ctx);
 
-int nvglCreateImageFromHandleGLES2(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
-GLuint nvglImageHandleGLES2(NVGcontext* ctx, int image);
+static int nvglCreateImageFromHandleGLES2(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
+static GLuint nvglImageHandleGLES2(NVGcontext* ctx, int image);
 
 #endif
 
 #if defined NANOVG_GLES3
 
-NVGcontext* nvgCreateGLES3(int flags);
-void nvgDeleteGLES3(NVGcontext* ctx);
+static NVGcontext* nvgCreateGLES3(int flags);
+static void nvgDeleteGLES3(NVGcontext* ctx);
 
-int nvglCreateImageFromHandleGLES3(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
-GLuint nvglImageHandleGLES3(NVGcontext* ctx, int image);
+static int nvglCreateImageFromHandleGLES3(NVGcontext* ctx, GLuint textureId, int w, int h, int flags);
+static GLuint nvglImageHandleGLES3(NVGcontext* ctx, int image);
 
 #endif
 
@@ -1620,13 +1620,13 @@ static void glnvg__renderDelete(void* uptr) {
 }
 
 #if defined NANOVG_GL2
-NVGcontext* nvgCreateGL2(int flags)
+static NVGcontext* nvgCreateGL2(int flags)
 #elif defined NANOVG_GL3
-NVGcontext* nvgCreateGL3(int flags)
+static NVGcontext* nvgCreateGL3(int flags)
 #elif defined NANOVG_GLES2
-NVGcontext* nvgCreateGLES2(int flags)
+static NVGcontext* nvgCreateGLES2(int flags)
 #elif defined NANOVG_GLES3
-NVGcontext* nvgCreateGLES3(int flags)
+static NVGcontext* nvgCreateGLES3(int flags)
 #endif
 {
   NVGparams params;
@@ -1666,26 +1666,26 @@ error:
 }
 
 #if defined NANOVG_GL2
-void nvgDeleteGL2(NVGcontext* ctx)
+static void nvgDeleteGL2(NVGcontext* ctx)
 #elif defined NANOVG_GL3
-void nvgDeleteGL3(NVGcontext* ctx)
+static void nvgDeleteGL3(NVGcontext* ctx)
 #elif defined NANOVG_GLES2
-void nvgDeleteGLES2(NVGcontext* ctx)
+static void nvgDeleteGLES2(NVGcontext* ctx)
 #elif defined NANOVG_GLES3
-void nvgDeleteGLES3(NVGcontext* ctx)
+static void nvgDeleteGLES3(NVGcontext* ctx)
 #endif
 {
   nvgDeleteInternal(ctx);
 }
 
 #if defined NANOVG_GL2
-int nvglCreateImageFromHandleGL2(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
+static int nvglCreateImageFromHandleGL2(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
 #elif defined NANOVG_GL3
-int nvglCreateImageFromHandleGL3(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
+static int nvglCreateImageFromHandleGL3(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
 #elif defined NANOVG_GLES2
-int nvglCreateImageFromHandleGLES2(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
+static int nvglCreateImageFromHandleGLES2(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
 #elif defined NANOVG_GLES3
-int nvglCreateImageFromHandleGLES3(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
+static int nvglCreateImageFromHandleGLES3(NVGcontext* ctx, GLuint textureId, int w, int h, int imageFlags)
 #endif
 {
   GLNVGcontext* gl = (GLNVGcontext*)nvgInternalParams(ctx)->userPtr;
@@ -1703,13 +1703,13 @@ int nvglCreateImageFromHandleGLES3(NVGcontext* ctx, GLuint textureId, int w, int
 }
 
 #if defined NANOVG_GL2
-GLuint nvglImageHandleGL2(NVGcontext* ctx, int image)
+static GLuint nvglImageHandleGL2(NVGcontext* ctx, int image)
 #elif defined NANOVG_GL3
-GLuint nvglImageHandleGL3(NVGcontext* ctx, int image)
+static GLuint nvglImageHandleGL3(NVGcontext* ctx, int image)
 #elif defined NANOVG_GLES2
-GLuint nvglImageHandleGLES2(NVGcontext* ctx, int image)
+static GLuint nvglImageHandleGLES2(NVGcontext* ctx, int image)
 #elif defined NANOVG_GLES3
-GLuint nvglImageHandleGLES3(NVGcontext* ctx, int image)
+static GLuint nvglImageHandleGLES3(NVGcontext* ctx, int image)
 #endif
 {
   GLNVGcontext* gl = (GLNVGcontext*)nvgInternalParams(ctx)->userPtr;